Create a List of Evaluation Metrics

This function creates a data frame containing names of evaluation metrics, their types, and the total count of metrics. Each metric is categorized by its type, which is represented as an integer.

measures()

Value

A data frame with the following components:

names: A character vector of metric names.

type: An integer vector representing the type of each metric.

The total number of metrics is equal to the length of the names vector.

Details

The type vector categorizes metrics as follows:

1: Metrics that are generally used for performance evaluation.

0: Metrics that are less commonly used or have different evaluation criteria.

Examples

# Create the data frame of metrics
metric_df <- measures()

# Print the data frame
print(metric_df)
#>                names type
#> 1           accuracy    1
#> 2  average-precision    1
#> 3                clp    0
#> 4           coverage    0
#> 5                 f1    1
#> 6       hamming-loss    0
#> 7          macro-auc    1
#> 8        macro-auprc    1
#> 9           macro-f1    1
#> 10   macro-precision    1
#> 11      macro-recall    1
#> 12       margin-loss    0
#> 13         micro-auc    1
#> 14       micro-auprc    1
#> 15          micro-f1    1
#> 16   micro-precision    1
#> 17      micro-recall    1
#> 18               mlp    0
#> 19         one-error    0
#> 20         precision    1
#> 21      ranking-loss    0
#> 22            recall    1
#> 23           roc-auc    1
#> 24     roc-auc-macro    1
#> 25     roc-auc-micro    1
#> 26   subset-accuracy    1
#> 27               wlp    0