Create a List of Evaluation Metrics

This function creates a data frame containing the names of evaluation metrics, their corresponding types, and the total count of metrics. Each metric is categorized by its type, which indicates whether a higher or lower value represents better performance.

wtl.measures()

Value

A data frame with the following columns:

names: A character vector of metric names.

type: An integer vector representing the evaluation type of each metric.

The total number of metrics equals the length of the names vector.

Details

The type vector categorizes metrics as follows:

1: Metrics where a higher value indicates better performance.

0: Metrics where a lower value indicates better performance.

Below is a summary of the metrics and their evaluation types:

Metric Name	Type	Interpretation

accuracy	1	higher is better

auprc_macro	1	higher is better

auprc_micro	1	higher is better

auprc_samples	1	higher is better

auprc_weighted	1	higher is better

average_precision	1	higher is better

clp	0	lower is better

coverage	0	lower is better

f1	1	higher is better

hamming_loss	0	lower is better

macro_auc	1	higher is better

macro_F1	1	higher is better

macro_precision	1	higher is better

macro_recall	1	higher is better

margin_loss	0	lower is better

micro_auc	1	higher is better

micro_F1	1	higher is better

micro_precision	1	higher is better

micro_recall	1	higher is better

mlp	0	lower is better

one_error	0	lower is better

precision	1	higher is better

ranking_loss	0	lower is better

recall	1	higher is better

roc_auc_macro	1	higher is better

roc_auc_micro	1	higher is better

roc_auc_samples	1	higher is better

roc_auc_weighted	1	higher is better

subset_accuracy	1	higher is better

wlp	0	lower is better

Examples

# Create the data frame of metrics
metric_df <- wtl.measures()

# Display the resulting data frame
print(metric_df)
#>                names type
#> 1           accuracy    1
#> 2        auprc_macro    1
#> 3        auprc_micro    1
#> 4      auprc_samples    1
#> 5     auprc_weighted    1
#> 6  average_precision    1
#> 7                clp    0
#> 8           coverage    0
#> 9                 f1    1
#> 10      hamming_loss    0
#> 11         macro_auc    1
#> 12          macro_f1    1
#> 13   macro_precision    1
#> 14      macro_recall    1
#> 15       margin_loss    0
#> 16         micro_auc    1
#> 17          micro_f1    1
#> 18   micro_precision    1
#> 19      micro_recall    1
#> 20               mlp    0
#> 21         one_error    0
#> 22         precision    1
#> 23      ranking_loss    0
#> 24            recall    1
#> 25     roc_auc_macro    1
#> 26     roc_auc_micro    1
#> 27   roc_auc_samples    1
#> 28  roc_auc_weighted    1
#> 29   subset_accuracy    1
#> 30               wlp    0