Compute Win-Tie-Loss Summary for Methods

This function calculates the win, tie, and loss summary between all pairs of methods based on their scores for a given measure type.

win.tie.loss.compute(data, measure.type)

Arguments

data

A data frame containing the scores of different methods (columns are methods, rows are observations). NA values are replaced by 0.

measure.type

An integer indicating the measure type:

1: Higher scores are better.

0: Lower scores are better.

Value

A data frame summarizing the number of wins, ties, and losses for each method. The returned data frame has columns:

method: Method name.

win: Total number of wins for this method.

tie: Total number of ties for this method.

loss: Total number of losses for this method.

Details

For each pair of methods, it compares their scores: - If `measure.type = 1`: higher scores are considered better (win if method1 > method2) - If `measure.type = 0`: lower scores are considered better (win if method1 < method2)

Examples

if (FALSE) { # \dontrun{
# Load example data
name.file = "~/WinTieLoss/Data/random-data.csv"
data = data.frame(read.csv(name.file))
data = data[,-1]  # Remove first column if needed
methods.names = colnames(data)

# Load metric types
df_res.mes <- wtl.measures()
filtered_res.mes <- dplyr::filter(df_res.mes, names == "clp")
measure.type = as.numeric(filtered_res.mes$type)

# Compute Win-Tie-Loss summary
res = win.tie.loss.compute(data = data, measure.type = measure.type)
res$method <- factor(res$method, levels = methods.names)
res <- res[order(res$method), ]

# Save results to CSV
FolderResults <- "~/WinTieLoss/Results"
save = paste(FolderResults, "/clp.csv", sep = "")
write.csv(res, save, row.names = FALSE)

# Plot results (optional)
wtl = c("win", "tie", "loss")
colnames(res) = wtl 
save = paste(FolderResults, "/clp.pdf", sep = "")
win.tie.loss.plot(
  data = res, 
  names.methods = methods.names, 
  name.file = save, 
  width = 18, height = 10, 
  bottom = 2, left = 11, top = 0, right = 1, 
  size.font = 2.0, wtl = wtl
)
} # }